add min=1 all'input type number

+++ database/migrations/2023_09_19_172905_create_books_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
             $table->integer('pages')->nullable()->default(1);
             $table->mediumText('description')->nullable();
             $table->string('year')->nullable();
-            $table->integer('price');
+            $table->float('price');
             $table->timestamps();
         });
     }
